+++ app/uk/gov/hmrc/selfassessmentapi/domain/TaxYearPropertyTypes.scala
@@ -19,9 +19,10 @@ package uk.gov.hmrc.selfassessmentapi.domain
 import uk.gov.hmrc.selfassessmentapi.domain.blindperson.BlindPersons
 import uk.gov.hmrc.selfassessmentapi.domain.charitablegiving.CharitableGivings
 import uk.gov.hmrc.selfassessmentapi.domain.pensioncontribution.PensionContributions
+import uk.gov.hmrc.selfassessmentapi.domain.taxrefundedorsetoff.TaxRefundedOrSetOffs
 
 object TaxYearPropertyTypes {
-  val types = Seq(PensionContributions, CharitableGivings, BlindPersons)
+  val types = Seq(PensionContributions, CharitableGivings, BlindPersons, TaxRefundedOrSetOffs)
   private val typesByName = types.map(x => x.name -> x).toMap
   def fromName(name: String): Option[TaxYearPropertyType] = typesByName.get(name)
 }